refactor: the Header component has been changed
