+++ packages/client/src/components/Navigation/Navigation.module.scss
@@ -1,62 +1,41 @@
 .nav {
   display: flex;
+  justify-content: center;
+  background-color: #2D3142;
+}
+
+.wrapper {
+  flex: 1;
+  display: flex;
   align-items: center;
   justify-content: space-between;
-  padding: 36px 48px;
-  background-color: white;
-
-  .logo {
-    font-size: 24px;
-    font-style: normal;
-    font-weight: 900;
-    line-height: normal;
-    letter-spacing: 0.72px;
-    text-transform: uppercase;
-  }
+  padding: 22px 50px;
+  max-width: 1024px;
 
   ul {
     display: flex;
     align-items: center;
-    gap: 35px;
+    column-gap: 40px;
 
     li {
       display: flex;
+      column-gap: 7px;
       align-items: center;
       list-style-type: none;
     }
-
-    a {
-      display: inline-block;
-      font-size: 24px;
-      font-style: normal;
-      font-weight: 900;
-      line-height: normal;
-      letter-spacing: 0.72px;
-      text-transform: uppercase;
-    }
   }
 }
-.active {
-  background: var(--link-activek);
-  border-radius: 128px;
-  padding: 12px 24px;
-}
 
-.btnExit {
-  width: fit-content;
-  display: flex;
-  align-items: center;
-  margin: 0;
-  padding: 0;
-  background-color: transparent;
-  color: #1F252D;
-  font-size: 24px;
+.link {
+  font-size: 15px;
   font-style: normal;
   font-weight: 900;
   line-height: normal;
-  letter-spacing: 0.72px;
+  letter-spacing: 0.45px;
   text-transform: uppercase;
-  svg {
-    margin-right: 10px;
-  }
+  color: #fff;
+}
+
+.active  {
+
 }
