--- packages/client/src/config/navConfig.ts
@@ -1,24 +1,0 @@
-import { AppPath } from '@/types/AppPath'
-
-export const navConfig = [
-  {
-    text: 'О проекте',
-    path: AppPath.MAIN,
-  },
-  {
-    text: 'Правила игры',
-    path: AppPath.RULES,
-  },
-  {
-    text: 'Форум',
-    path: AppPath.FORUM,
-  },
-  {
-    text: 'Таблица лидеров',
-    path: AppPath.LEADERBOARD,
-  },
-  {
-    text: 'Профиль',
-    path: AppPath.PROFILE,
-  },
-]
